Consistent checking for success of filestream open.
Previously, filestreams were not checked for failure state
after opening a file, which lead to undefined behaviour if an
opened file did not exist: Feeding such a failure-state stream
into Circuit lead to overallocation of memory and crashes (e.g.).

Now all filestream open calls are checked for success and exit
immediately with an error message if the stream is in failure state.